feat: buy uniswap button in action card
