+++ apps/frontend/acquisition-royale-home/src/features/mvp/trade/index.ts
@@ -1,0 +1,3 @@
+import Trade from './Trade'
+
+export default Trade
